backend/pyproject.toml
[project]
name = "singapore-smb-backend"
version = "0.1.0"
description = "Django backend for Singapore SMB E-commerce Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=6.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.4.0", # JWT Authentication
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-allauth>=65.13.1",
    "django-debug-toolbar>=4.5.0", # Development debugging
    "celery>=5.5.3",
    "redis>=6.4.0",
    "psycopg2-binary>=2.9.10",
    "gunicorn>=21.2.0",
    "pillow>=11.3.0",
    "drf-spectacular>=0.27.1", # For API Documentation
    "stripe>=8.4.0", # Payment Processing
    "weasyprint>=67.0", # Invoice PDF Generation
    "python-jose>=3.5.0", # JWT Handling # Error Monitoring
    "sentry-sdk>=2.37.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "factory-boy>=3.3.3",
    "django-filter>=25.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "pytest-cov>=6.2.0",
    "black>=25.9.0",
    "isort>=7.0.0",
    "flake8>=7.3.0",
    "factory-boy>=3.3.3"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "apps"]
